admin and employee view done

+++ RestaurantOldies/model/Item.cs
@@ -10,17 +10,17 @@ namespace RestaurantOldies.model
     {
         public int id { get; set; }
         public string name { get; set; }
-        public int price { get; set; }
+        public float price { get; set; }
         public int quantity { get; set; }
 
-        public Item(string name, int price, int quantity)
+        public Item(string name, float price, int quantity)
         {
             this.name = name;
             this.price = price;
             this.quantity = quantity;
         }
 
-        public Item(int id, string name, int price, int quantity)
+        public Item(int id, string name, float price, int quantity)
         {
             this.id = id;
             this.name = name;
